kv: rename observedts package to uncertainty
This commit renames the `kvserver/observedts` package to `kvserver/uncertainty`,
in preparation for its role increasing in scope.
