+++ pkg/BUILD.bazel
@@ -134,7 +134,6 @@ ALL_TESTS = [
     "//pkg/kv/kvserver/idalloc:idalloc_test",
     "//pkg/kv/kvserver/intentresolver:intentresolver_test",
     "//pkg/kv/kvserver/liveness:liveness_test",
-    "//pkg/kv/kvserver/observedts:observedts_test",
     "//pkg/kv/kvserver/protectedts/ptcache:ptcache_test",
     "//pkg/kv/kvserver/protectedts/ptreconcile:ptreconcile_test",
     "//pkg/kv/kvserver/protectedts/ptstorage:ptstorage_test",
@@ -152,6 +151,8 @@ ALL_TESTS = [
     "//pkg/kv/kvserver/tscache:tscache_test",
     "//pkg/kv/kvserver/txnrecovery:txnrecovery_test",
     "//pkg/kv/kvserver/txnwait:txnwait_test",
+    "//pkg/kv/kvserver/uncertainty:observedts_test",
+    "//pkg/kv/kvserver/uncertainty:uncertainty_test",
     "//pkg/kv/kvserver:kvserver_test",
     "//pkg/kv:kv_test",
     "//pkg/migration/migrationcluster:migrationcluster_test",
