+++ pkg/kv/kvserver/BUILD.bazel
@@ -127,7 +127,6 @@ go_library(
         "//pkg/kv/kvserver/kvserverpb",
         "//pkg/kv/kvserver/liveness",
         "//pkg/kv/kvserver/liveness/livenesspb",
-        "//pkg/kv/kvserver/observedts",
         "//pkg/kv/kvserver/protectedts",
         "//pkg/kv/kvserver/protectedts/ptpb:ptpb_go_proto",
         "//pkg/kv/kvserver/raftentry",
@@ -142,6 +141,7 @@ go_library(
         "//pkg/kv/kvserver/tscache",
         "//pkg/kv/kvserver/txnrecovery",
         "//pkg/kv/kvserver/txnwait",
+        "//pkg/kv/kvserver/uncertainty",
         "//pkg/multitenant",
         "//pkg/multitenant/tenantcostmodel",
         "//pkg/roachpb:with-mocks",
