+++ pkg/kv/kvserver/uncertainty/BUILD.bazel
@@ -32,3 +32,35 @@ go_test(
         "@com_github_stretchr_testify//require",
     ],
 )
+
+go_library(
+    name = "uncertainty",
+    srcs = [
+        "doc.go",
+        "limit.go",
+        "uncertainty.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/uncertainty",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/kv/kvserver/kvserverpb",
+        "//pkg/roachpb:with-mocks",
+        "//pkg/util/hlc",
+    ],
+)
+
+go_test(
+    name = "uncertainty_test",
+    srcs = [
+        "limit_test.go",
+        "uncertainty_test.go",
+    ],
+    embed = [":uncertainty"],
+    deps = [
+        "//pkg/kv/kvserver/kvserverpb",
+        "//pkg/roachpb:with-mocks",
+        "//pkg/util/hlc",
+        "//pkg/util/leaktest",
+        "@com_github_stretchr_testify//require",
+    ],
+)
