+++ pkg/kv/kvserver/uncertainty/limit_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package observedts
+package uncertainty
 
 import (
 	"testing"
